♻️ Refact: APIProviderStore 클래스 Store 폴더 아래로 이동
